requirements.txt
flask
langchain
langchain-openai
langchain-community
langgraph
openai
python-dotenv